refactor: extract sessionStorage persistence from OnboardingWizard.tsx
Closes #978 partially.

Pulls the wizard's sessionStorage checkpoint layer + the shared
types it owns into wizard/persistence.ts:

- WIZARD_STATE_KEY constant
- PersistedWizardState, StepHistoryEntry, StackSubStep, WizardStep
  type definitions
- normalizeStepHistory legacy-migration helper
- loadPersistedWizardState, clearPersistedWizardState
- new savePersistedWizardState helper for the on-change effect

OnboardingWizard.tsx drops from 1497 to 1430 LOC. Zero behaviour
change — pure module move, no React in the new file.

The "state machine" half of the suggested seam in the issue
(setCurrentStep + stepHistory + navigateTo + handleBack) and the
useExpressInstall hook stay open for a follow-up. Both reach into a
broad set of in-component React state, and the issue's own warnings
about deploy-critical-path + every-operator-sees-this risk justify
landing them only with a dedicated PR + a per-stage live FCoS
validation, which this commit deliberately does not attempt.
